IFEDMethodBase, part 9.
